feat: creates use-cases for each service method
