+++ server/src/auth/use-cases/verifyRefreshToken/verify-refresh-token.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@nestjs/common';
+
+import { RefreshTokenRepository } from '@auth/infra/typeorm/repositories';
+
+@Injectable()
+export class VerifyRefreshToken {
+  constructor(private refreshTokenRepository: RefreshTokenRepository) {}
+
+  async execute(token: string): Promise<boolean> {
+    const refreshToken = await this.refreshTokenRepository.findRefreshToken(
+      token,
+    );
+
+    return !!refreshToken;
+  }
+}
